requirements.txt
email-validator
Flask
Flask-Babel
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
PyJWT
python-dotenv